+++ src/neural-style-transfer/transport.go
@@ -5,14 +5,14 @@ import (
 	"encoding/base64"
 	"encoding/json"
 	"errors"
-	"fmt"
 	"html/template"
 	"net/http"
 	"path/filepath"
 	"strconv"
 	"sync"
 
-	"github.com/go-kit/kit/log"
+	"neural-style-util"
+
 	httptransport "github.com/go-kit/kit/transport/http"
 	"github.com/gorilla/mux"
 )
@@ -21,54 +21,6 @@ var (
 	// ErrBadRouting define the default routing error information
 	ErrBadRouting = errors.New("inconsistent mapping between route and handler (programmer error)")
 )
-
-func getUsername(req *http.Request) string {
-	var authString = ""
-	if len(req.Header) > 0 {
-		for k, v := range req.Header {
-			if k == "Authorization" {
-				authString = v[0]
-			}
-		}
-	}
-
-	if authString == "" {
-		return ""
-	}
-
-	_, username := CheckToken(authString)
-	return username
-}
-
-func usernameMiddleware(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		getUsername(r)
-		next.ServeHTTP(w, r)
-	})
-}
-
-func authMiddleware(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		username := getUsername(r)
-		if username == "" {
-			//w.Header().Set("Content-Type", "text/html; text/javascript; text/css; charset=utf-8")
-			w.WriteHeader(http.StatusUnauthorized)
-			w.Write([]byte("Unauthorized"))
-		} else {
-			fmt.Println("user name is ", username)
-			next.ServeHTTP(w, r)
-		}
-	})
-}
-
-func accessControl(h http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Access-Control-Allow-Origin", "*")
-		w.Header().Set("Access-Control-Allow-Methods", "GET, POST, OPTIONS")
-
-		h.ServeHTTP(w, r)
-	})
-}
 
 // templ represents a single template
 type templateHandler struct {
@@ -87,122 +39,22 @@ func (t *templateHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 }
 
 // MakeHTTPHandler generate the http handler for the style service handler
-func MakeHTTPHandler(ctx context.Context, endpoint Endpoints, logger log.Logger) http.Handler {
-	r := mux.NewRouter()
-	options := []httptransport.ServerOption{
-		httptransport.ServerErrorLogger(logger),
-		httptransport.ServerErrorEncoder(encodeError),
-	}
-
+func MakeHTTPHandler(ctx context.Context, r *mux.Router, svc Service, options ...httptransport.ServerOption) *mux.Router {
 	//GET /styleTransfer/{content}/{style}/{iterations}
-	r.Methods("GET").Path("/styleTransfer").Queries("content", "{content}", "style", "{style}", "iterations", "{iterations:[0-9]+}").Handler(authMiddleware(httptransport.NewServer(
-		endpoint.NSEndpoint,
+	r.Methods("GET").Path("/styleTransfer").Queries("content", "{content}", "style", "{style}", "iterations", "{iterations:[0-9]+}").Handler(NSUtil.AuthMiddleware(httptransport.NewServer(
+		MakeNSEndpoint(svc),
 		decodeNSRequest,
 		encodeNSResponse,
 		options...,
 	)))
 
 	//GET /styleTransferPreview/{content}/{style}
-	r.Methods("GET").Path("/styleTransferPreview").Queries("content", "{content}", "style", "{style}").Handler(authMiddleware(httptransport.NewServer(
-		endpoint.NSPreviewEndpoint,
+	r.Methods("GET").Path("/styleTransferPreview").Queries("content", "{content}", "style", "{style}").Handler(NSUtil.AuthMiddleware(httptransport.NewServer(
+		MakeNSPreviewEndpoint(svc),
 		decodeNSPreviewRequest,
 		encodeNSResponse,
 		options...,
 	)))
-
-	// POST /styleTransfer/content
-	contentUploadHandler := httptransport.NewServer(
-		endpoint.NSContentUploadEndpoint,
-		decodeNSUploadContentRequest,
-		encodeNSUploadContentResponse,
-		options...,
-	)
-	r.Methods("POST").Path("/api/upload/content").Handler(authMiddleware(accessControl(contentUploadHandler)))
-
-	// POST /styleTransfer/style
-	styleUploadHandler := httptransport.NewServer(
-		endpoint.NSStyleUploadEndpoint,
-		decodeNSUploadStyleRequest,
-		encodeNSUploadStyleResponse,
-		options...,
-	)
-	r.Methods("POST").Path("/api/upload/style").Handler(authMiddleware(accessControl(styleUploadHandler)))
-
-	// GET /api/artists
-	r.Methods("GET").Path("/api/artists").Handler(authMiddleware(accessControl(httptransport.NewServer(
-		endpoint.NSGetArtistsEndpoint,
-		decodeNSGetArtistRequest,
-		encodeNSGetArtistsResponse,
-		options...,
-	))))
-
-	// GET /api/artists/hotest
-	r.Methods("GET").Path("/api/artists/hotest").Handler(authMiddleware(accessControl(httptransport.NewServer(
-		endpoint.NSGetHotestArtistsEndpoint,
-		decodeNSGetArtistRequest,
-		encodeNSGetArtistsResponse,
-		options...,
-	))))
-
-	// Register
-	registerHandler := httptransport.NewServer(
-		endpoint.NSRegisterEndpoint,
-		decodeNSRegisterRequest,
-		encodeNSRegisterResponse,
-		options...,
-	)
-	r.Methods("POST").Path("/api/register").Handler(accessControl(registerHandler))
-
-	// Login
-	loginHandler := httptransport.NewServer(
-		endpoint.NSLoginEndpoint,
-		decodeNSLoginRequest,
-		encodeNSLoginResponse,
-		options...,
-	)
-	r.Methods("POST").Path("/api/authenticate").Handler(accessControl(loginHandler))
-
-	// GET api/products
-	r.Methods("GET").Path("/api/products").Handler(usernameMiddleware(httptransport.NewServer(
-		endpoint.NSGetProductsEndpoint,
-		decodeNSGetProductsRequest,
-		encodeNSGetProductsResponse,
-		options...,
-	)))
-
-	// GET api/products/{id}
-	r.Methods("GET").Path("/api/products/{id}").Handler(httptransport.NewServer(
-		endpoint.NSGetProductsByIDEndpoint,
-		decodeNSGetProductByIDRequest,
-		encodeNSGetProductByIdResponse,
-		options...,
-	))
-
-	// GET api/products/{id}/reviews
-	r.Methods("GET").Path("/api/products/{id}/reviews").Handler(authMiddleware(httptransport.NewServer(
-		endpoint.NSGetReviewsByIDEndpoint,
-		decodeNSGetReviewsByIDRequest,
-		encodeNSGetReviewsByIDResponse,
-		options...,
-	)))
-
-	// output file server
-	outputFiles := http.FileServer(http.Dir("data/outputs/"))
-	r.PathPrefix("/outputs/").Handler(http.StripPrefix("/outputs/", outputFiles))
-
-	// style file server
-	styleFiles := http.FileServer(http.Dir("data/styles/"))
-	r.PathPrefix("/styles/").Handler(http.StripPrefix("/styles", styleFiles))
-
-	// content file server
-	contentFiles := http.FileServer(http.Dir("data/contents"))
-	r.PathPrefix("/contents/").Handler(http.StripPrefix("/contents/", contentFiles))
-
-	// template file
-	resourceFile := http.FileServer(http.Dir("dist"))
-	r.PathPrefix("/static/").Handler(http.StripPrefix("/static/", resourceFile))
-
-	r.Path("/").Handler(resourceFile)
 
 	return r
 }
@@ -265,100 +117,6 @@ func decodeNeuralStyleCommonParams(vars map[string]string) (string, string, erro
 	return string(contentPath), string(stylePath), nil
 }
 
-func decodeNSUploadContentRequest(_ context.Context, r *http.Request) (interface{}, error) {
-	productData := Product{ID: "1"}
-	json.NewDecoder(r.Body).Decode(&productData)
-	return NSUploadRequest{ProductData: productData}, nil
-}
-
-func encodeNSUploadContentResponse(ctx context.Context, w http.ResponseWriter, response interface{}) error {
-	contentRes := response.(NSGetProductResponse)
-	if contentRes.Err != nil {
-		return contentRes.Err
-	}
-
-	w.Header().Set("context-type", "application/json, charset=utf8")
-	return json.NewEncoder(w).Encode(contentRes.Target)
-}
-
-func decodeNSUploadStyleRequest(_ context.Context, r *http.Request) (interface{}, error) {
-	productData := Product{ID: "1"}
-	json.NewDecoder(r.Body).Decode(&productData)
-	return NSUploadRequest{ProductData: productData}, nil
-}
-
-func encodeNSUploadStyleResponse(ctx context.Context, w http.ResponseWriter, response interface{}) error {
-	styleRes := response.(NSGetProductResponse)
-	if styleRes.Err != nil {
-		return styleRes.Err
-	}
-
-	w.Header().Set("context-type", "application/json, charset=utf8")
-	return json.NewEncoder(w).Encode(styleRes.Target)
-}
-
-func decodeNSGetProductsRequest(_ context.Context, r *http.Request) (interface{}, error) {
-	return nil, nil
-}
-
-func encodeNSGetProductsResponse(ctx context.Context, w http.ResponseWriter, response interface{}) error {
-	productsRes := response.(NSGetProductsResponse)
-	if productsRes.Err != nil {
-		return productsRes.Err
-	}
-
-	w.Header().Set("context-type", "application/json, charset=utf8")
-	return json.NewEncoder(w).Encode(productsRes.Products)
-}
-
-func decodeNSGetArtistRequest(_ context.Context, r *http.Request) (interface{}, error) {
-	return nil, nil
-}
-
-func encodeNSGetArtistsResponse(ctx context.Context, w http.ResponseWriter, res interface{}) error {
-	artistsRes := res.(NSGetArtistsResponse)
-	if artistsRes.Err != nil {
-		return artistsRes.Err
-	}
-
-	w.Header().Set("content-type", "application/json, charset=utf8")
-	return json.NewEncoder(w).Encode(artistsRes.Artists)
-}
-
-func decodeNSGetProductByIDRequest(_ context.Context, r *http.Request) (interface{}, error) {
-	vars := mux.Vars(r)
-	id := vars["id"]
-
-	return NSGetProductByIDRequest{ID: id}, nil
-}
-
-func encodeNSGetProductByIdResponse(ctx context.Context, w http.ResponseWriter, response interface{}) error {
-	productRes := response.(NSGetProductResponse)
-	if productRes.Err != nil {
-		return productRes.Err
-	}
-
-	w.Header().Set("context-type", "application/json, charset=utf8")
-	return json.NewEncoder(w).Encode(productRes.Target)
-}
-
-func decodeNSGetReviewsByIDRequest(_ context.Context, r *http.Request) (interface{}, error) {
-	vars := mux.Vars(r)
-	id := vars["id"]
-
-	return NSGetReviewsByIDRequest{ID: id}, nil
-}
-
-func encodeNSGetReviewsByIDResponse(ctx context.Context, w http.ResponseWriter, response interface{}) error {
-	reviewsRes := response.(NSGetReviewsByIDResponse)
-	if reviewsRes.Err != nil {
-		return reviewsRes.Err
-	}
-
-	w.Header().Set("context-type", "application/json, charset=utf8")
-	return json.NewEncoder(w).Encode(reviewsRes.Reviews)
-}
-
 type errorer interface {
 	error() error
 }
@@ -384,36 +142,4 @@ func encodeError(ctx context.Context, err error, w http.ResponseWriter) {
 	json.NewEncoder(w).Encode(map[string]interface{}{
 		"error": err.Error(),
 	})
-}
-
-func decodeNSRegisterRequest(_ context.Context, r *http.Request) (interface{}, error) {
-	userData := UserInfo{ID: "1"}
-	json.NewDecoder(r.Body).Decode(&userData)
-	return NSAuthenticationRequest{UserData: userData}, nil
-}
-
-func encodeNSRegisterResponse(ctx context.Context, w http.ResponseWriter, response interface{}) error {
-	if e, ok := response.(errorer); ok && e.error() != nil {
-		encodeError(ctx, e.error(), w)
-		return nil
-	}
-
-	w.Header().Set("context-type", "application/json, charset=utf8")
-	return json.NewEncoder(w).Encode(response)
-}
-
-func decodeNSLoginRequest(_ context.Context, r *http.Request) (interface{}, error) {
-	userData := UserInfo{ID: "1"}
-	json.NewDecoder(r.Body).Decode(&userData)
-	return NSAuthenticationRequest{UserData: userData}, nil
-}
-
-func encodeNSLoginResponse(ctx context.Context, w http.ResponseWriter, response interface{}) error {
-	loginRes := response.(NSLoginResponse)
-	if loginRes.Err != nil {
-		return loginRes.Err
-	}
-
-	w.Header().Set("context-type", "application/json, charset=utf8")
-	return json.NewEncoder(w).Encode(loginRes.Target)
 }
