+++ src/neural-style-util/error.go
@@ -1,0 +1,1 @@
+package NSUtil
